feat(backend): AdminDashboardResponse dto implement

+++ backend/src/main/java/com/pankassi/backend/dto/response/Admin/AdminDashboardResponse.java
@@ -1,9 +1,9 @@
 package com.pankassi.backend.dto.response.Admin;
 
 public record AdminDashboardResponse(
-        Integer patientNumber,
-        Integer doctorNumber,
-        Integer staffNumber,
-        Integer waitingRequestNumber
+        long patientNumber,
+        long doctorNumber,
+        long staffNumber,
+        long waitingRequestNumber
 ) {
 }
